+++ src/transforms/kubernetes/mod.rs
@@ -1,3 +1,1 @@
-#[cfg(test)]
-mod kube_config;
 pub mod watch_client;
